+++ weave-js/src/components/PagePanelComponents/Home/Browse3/views/Leaderboard/query/leaderboardQuery.ts
@@ -3,17 +3,17 @@ import _ from 'lodash';
 
 import {flattenObjectPreservingWeaveTypes} from '../../../../Browse2/browse2Util';
 import {parseRefMaybe} from '../../../../Browse2/SmallRef';
-import {EVALUATE_OP_NAME_POST_PYDANTIC} from '../../common/heuristics';
-import {TraceServerClient} from '../../wfReactInterface/traceServerClient';
-import {TraceObjSchema} from '../../wfReactInterface/traceServerClientTypes';
+import {EVALUATE_OP_NAME_POST_PYDANTIC} from '../../../pages/common/heuristics';
+import {TraceServerClient} from '../../../pages/wfReactInterface/traceServerClient';
+import {TraceObjSchema} from '../../../pages/wfReactInterface/traceServerClientTypes';
 import {
   convertISOToDate,
   projectIdFromParts,
-} from '../../wfReactInterface/tsDataModelHooks';
+} from '../../../pages/wfReactInterface/tsDataModelHooks';
 import {
   objectVersionKeyToRefUri,
   opVersionKeyToRefUri,
-} from '../../wfReactInterface/utilities';
+} from '../../../pages/wfReactInterface/utilities';
 import {ALL_VALUE, FilterAndGroupSpec} from '../types/leaderboardConfigType';
 
 export type LeaderboardValueRecord = {
